web_requirements.txt
# Web UI requirements for crew feedback parser
Flask>=2.3.0
Werkzeug>=2.3.0

# Core dependencies (already in requirements.txt)
requests>=2.31.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
urllib3>=2.0.0
certifi>=2023.7.22